requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
gunicorn==23.0.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]
python-dotenv==1.0.1
httpx==0.27.2
redis==5.0.8
boto3==1.35.36
greenstalk

